requirements.txt
discord.py
requests
pandas
pandas_ta
python-dotenv
